Move all the lingui junk together
There's a bunch of crap we're adding in to get lingui to work in a
reasonable way in our project.

Threw everything together into one file and so that at least we
can contain the blast radius of this stuff.
