+++ web/src/components/Footer.js
@@ -3,7 +3,7 @@ import withContext from '../withContext'
 import { contextPropTypes } from '../context'
 import PropTypes from 'prop-types'
 import { Trans, withI18n } from 'lingui-react'
-import { translateText } from '../utils/initLingui'
+import { translateText } from '../utils/linguiUtils'
 import { WordMark } from '@cdssnc/gcui'
 import styled, { css } from 'react-emotion'
 import { theme, mediaQuery, visuallyhiddenMobile } from '../styles'
